requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-dotenv==1.0.0
openai==1.54.3
qdrant-client==1.6.9
sentence-transformers==3.0.1
requests==2.31.0
tqdm==4.66.1
numpy==1.24.3
python-multipart==0.0.6
torch>=1.9.0
transformers>=4.21.0
scikit-learn>=1.0.0